refactor(v2): drop -dev suffix, clean up headers, and additional HW envs
- add envs support for HW1.0 and 1.1
- state.h folds into pocketpd.h and events.h
- move v1 to src/v1/main.cpp
