+++ include/v2/input/button_gesture.h
@@ -10,7 +10,7 @@
 #include <cstdint>
 #include <optional>
 
-#include "v2/state.h"
+#include "v2/pocketpd.h"
 
 namespace pocketpd {
