+++ include/v2/pocketpd.h
@@ -1,6 +1,6 @@
 /**
  * @file pocketpd.h
- * @brief PocketPD v2 product-level types and constants.
+ * @brief PocketPD v2 product-level types, configuration constants, etc.
  *
  * Header is kept dependency-free so it can be included from hardware adapters, stages, tasks, and
  * native tests alike.
@@ -53,5 +53,49 @@ namespace pocketpd {
      * bypasses the debounce.
      */
     constexpr uint32_t EEPROM_SAVE_DEBOUNCE_MS = 2000;
+
+    // —— Input types
+
+    /**
+     * @brief Identifier for the three physical buttons plus a synthetic L_R combo.
+     * `L` and `R` denote the left and right side buttons; `ENCODER` is the rotary push.
+     */
+    enum class ButtonId : uint8_t {
+        ENCODER,
+        R,
+        L,
+        L_R,
+    };
+
+    inline const char* to_string(ButtonId btn_id) {
+        switch (btn_id) {
+        case ButtonId::ENCODER:
+            return "ENCODER_BTN";
+        case ButtonId::L:
+            return "L_BTN";
+        case ButtonId::R:
+            return "R_BTN";
+        case ButtonId::L_R:
+            return "L_R_SYNTHETIC";
+        default:
+            return "UNKNOWN";
+        }
+    }
+
+    enum class Gesture : uint8_t {
+        SHORT,
+        LONG,
+    };
+
+    inline const char* to_string(Gesture g) {
+        switch (g) {
+        case Gesture::SHORT:
+            return "SHORT";
+        case Gesture::LONG:
+            return "LONG";
+        default:
+            return "UNKNOWN";
+        }
+    }
 
 } // namespace pocketpd
