+++ include/v2/stages/energy/energy_view.h
@@ -11,8 +11,9 @@
 #include <cstdint>
 #include <cstdio>
 
+#include "v2/events.h"
 #include "v2/images.h"
-#include "v2/state.h"
+#include "v2/pocketpd.h"
 
 namespace pocketpd {
