+++ include/v2/stages/normal/pps_mode.h
@@ -11,7 +11,6 @@
 #include "v2/events.h"
 #include "v2/hal/pd_sink_controller.h"
 #include "v2/pocketpd.h"
-#include "v2/state.h"
 
 namespace pocketpd {
