+++ include/v2/tasks/sensor_task.h
@@ -10,7 +10,7 @@
 #include "v2/app.h"
 #include "v2/events.h"
 #include "v2/hal/power_monitor.h"
-#include "v2/state.h"
+#include "v2/pocketpd.h"
 
 namespace pocketpd {
