+++ platformio.ini
@@ -34,6 +34,17 @@ lib_deps =
 	mathertel/RotaryEncoder@^1.5.3
 	khoih-prog/RPI_PICO_TimerInterrupt@^1.3.1
 
+[env:HW1_0]
+extends = pico_base
+extra_scripts = pre:scripts/configure_fmt_header_only.py
+build_flags =
+	${pico_base.build_flags}
+	-DHW1_0
+	-DVERSION="\"1.0.0\""
+build_src_filter =
+	-<*>
+	+<v1/**/*.cpp>
+
 [env:HW1_3]
 extends = pico_base
 extra_scripts = pre:scripts/configure_fmt_header_only.py
@@ -41,10 +52,39 @@ build_flags =
 	${pico_base.build_flags}
 	-DHW1_3
 	-DVERSION="\"1.0.0\""
-build_src_filter = 
-	-<*> 
-	+<main.cpp> 
+build_src_filter =
+	-<*>
 	+<v1/**/*.cpp>
+
+[env:HW1_0_V2]
+extends = pico_base
+extra_scripts =
+	pre:scripts/configure_fmt_header_only.py
+	post:scripts/compiledb_inject_headers.py
+build_flags =
+	${pico_base.build_flags}
+	-DHW_VERSION_MAJOR=1
+	-DHW_VERSION_MINOR=0
+	-DVERSION="\"2.0.0\""
+build_src_filter =
+	-<*>
+	+<main.cpp>
+	+<v2/**/*.cpp>
+
+[env:HW1_1_V2]
+extends = pico_base
+extra_scripts =
+	pre:scripts/configure_fmt_header_only.py
+	post:scripts/compiledb_inject_headers.py
+build_flags =
+	${pico_base.build_flags}
+	-DHW_VERSION_MAJOR=1
+	-DHW_VERSION_MINOR=1
+	-DVERSION="\"2.0.0\""
+build_src_filter =
+	-<*>
+	+<main.cpp>
+	+<v2/**/*.cpp>
 
 [env:HW1_3_V2]
 extends = pico_base
@@ -53,8 +93,9 @@ extra_scripts =
 	post:scripts/compiledb_inject_headers.py
 build_flags =
 	${pico_base.build_flags}
-	-DHW1_3_V2
-	-DVERSION="\"2.0.0-dev\""
+	-DHW_VERSION_MAJOR=1
+	-DHW_VERSION_MINOR=3
+	-DVERSION="\"2.0.0\""
 build_src_filter =
 	-<*>
 	+<main.cpp>
