+++ src/v1/main.cpp
@@ -1,0 +1,10 @@
+#include <Arduino.h>
+#include "v1/StateMachine.h"
+
+static StateMachine statemachine;
+
+void setup() {}
+
+void loop() {
+    statemachine.update();
+}
